created video.php theme template
just a crappy theme template page for custom post type "video" which used a custom field "youtube_id" containing the YouTube video ID. 
later, we'll use this template to show the featured image of the custom post type (for a placeholder image), we'll overlay the title and excerpt text, and provide the transcript as a collapsed text region the visitor can expand to read. when a visitor clicks, we can use fancybox or another popup view for the video playback.
for now, this just embeds the video so we can see/use it on the front end during content creation.